+++ context/app/static/js/components/workspaces/AddDatasetsTable/index.ts
@@ -1,3 +1,5 @@
 import AddDatasetsTable from './AddDatasetsTable';
+import { useDatasetsAutocomplete, UseDatasetsAutocompleteReturnType } from './hooks';
 
+export { useDatasetsAutocomplete, type UseDatasetsAutocompleteReturnType };
 export default AddDatasetsTable;
